fix: Move member's template into pages
